requirements.txt
asgiref==3.3.4
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Django==3.2.3
djangorestframework==3.12.4
Flask==2.0.1
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==3.0.0
urllib3==1.26.4
Werkzeug==2.0.1
django-cors-headers==3.11.0
openpyxl==3.0.9
djangorestframework-simplejwt==5.1.0
pyjwt==2.3.0
psycopg2==2.9.3
